requirements.txt
streamlit
google-generativeai
pandas
st-gsheets-connection